agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.35.0
httpx>=0.27.0
bedrock-agentcore>=0.1.0
pycryptodome>=3.20.0
coincurve>=20.0.0
